make base refactoring, split keeper into files
