+++ x/staking/src/keeper/bonded.rs
@@ -1,0 +1,65 @@
+pub use super::*;
+
+impl<
+        SK: StoreKey,
+        PSK: ParamsSubspaceKey,
+        AK: AccountKeeper<SK>,
+        BK: BankKeeper<SK>,
+        KH: KeeperHooks<SK>,
+    > Keeper<SK, PSK, AK, BK, KH>
+{
+    pub fn bonded_tokens_to_not_bonded<DB: Database, CTX: TransactionalContext<DB, SK>>(
+        &self,
+        ctx: &mut CTX,
+        amount: i64,
+    ) -> anyhow::Result<()> {
+        let params = self.staking_params_keeper.get(&ctx.multi_store())?;
+        let coins = SendCoins::new(vec![Coin {
+            denom: params.bond_denom,
+            amount: Uint256::from(amount as u64),
+        }])?;
+        Ok(self
+            .bank_keeper
+            .send_coins_from_module_to_module::<DB, AK, CTX>(
+                ctx,
+                BONDED_POOL_NAME.into(),
+                NOT_BONDED_POOL_NAME.into(),
+                coins,
+            )?)
+    }
+
+    pub fn bonded_to_unbonding<DB: Database, CTX: TransactionalContext<DB, SK>>(
+        &self,
+        ctx: &mut CTX,
+        validator: &mut Validator,
+    ) -> anyhow::Result<()> {
+        if validator.status != BondStatus::Bonded {
+            return Err(AppError::Custom(format!(
+                "bad state transition bonded to unbonding, validator: {:?}",
+                validator
+            ))
+            .into());
+        }
+        self.begin_unbonding_validator(ctx, validator)
+    }
+
+    pub fn bond_validator<DB: Database, CTX: TransactionalContext<DB, SK>>(
+        &self,
+        ctx: &mut CTX,
+        validator: &mut Validator,
+    ) -> anyhow::Result<()> {
+        // delete the validator by power index, as the key will change
+        self.delete_validator_by_power_index(ctx, validator);
+
+        validator.update_status(BondStatus::Bonded);
+        // save the now bonded validator record to the two referenced stores
+        self.set_validator(ctx, validator)?;
+        self.set_validator_by_power_index(ctx, validator)?;
+
+        // delete from queue if present
+        self.delete_validator_queue(ctx, validator)?;
+        // trigger hook
+        self.after_validator_bonded(ctx, validator)?;
+        Ok(())
+    }
+}
